# api/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "semrel registry API",
    "version": "1.0.0",
    "description": "The semrel plugin registry. The public catalogue is readable without authentication; publishing and review require a GitHub session.\n\nBrowsers authenticate with an HttpOnly `semrel_session` cookie set by the OAuth callback. API clients (the semrel CLI, CI jobs) send `Authorization: Bearer <token>`.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://registry.semrel.io",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/auth/github": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Start GitHub OAuth",
        "security": [],
        "parameters": [
          {
            "name": "next",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "A same-site path to return to. Absolute URLs are rejected."
          }
        ],
        "responses": {
          "307": {
            "description": "Redirect to GitHub."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "semrel_session",
        "description": "HttpOnly, SameSite=Lax session cookie set by the OAuth callback. Browser clients use this; mutating requests are additionally checked against the Origin header."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "For non-browser clients. The static ADMIN_TOKEN is accepted outside production only."
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Hub-Signature-256",
        "description": "HMAC-SHA256 over the raw request body, keyed with WEBHOOK_SECRET, formatted as sha256=<hex>."
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAuth": []
    }
  ]
}